+++ heima-leadnews/heima-leadnews-service/heima-leadnews-content/src/main/java/com/heima/content/mapper/achievement/ApAchievementMapper.java
@@ -1,0 +1,12 @@
+package com.heima.content.mapper.achievement;
+
+import com.baomidou.mybatisplus.core.mapper.BaseMapper;
+import com.heima.model.achievement.pojos.ApAchievement;
+import org.apache.ibatis.annotations.Mapper;
+
+/**
+ * 成就勋章定义表 Mapper
+ */
+@Mapper
+public interface ApAchievementMapper extends BaseMapper<ApAchievement> {
+}